requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
clarifai-grpc==10.0.10
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
idna==3.6
protobuf==4.25.2
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.4
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
